Prepare to apply metaclasses to generate new classes.
Adds metaclasses to the declaration specifier (conceptually part of the
type specifier), and to class declarations.

When the class definition is finished, the metaclass can be evaluated to
create a new class, which replaces the old one.

This probably does not work with forward declarations. We want to swap the
values instead of the old and new classes.