[Parse][Sema][Reflection] Fix metaclass declaration parsing
In Sema::CheckMetaclassName(), look up the metaclass name in the correct
declaration context without failing an assertion in
DeclContext::lookup() when the current declaration context is a
LinkageSpecDecl or an ExportDecl. This resolves #7.

Also, since the Invalid flag is initialized to ‘false’ in
default-constructed DeclResult objects, make
Sema::ActOnMetaclassDefinition() and Sema::CheckMetaclassName() return
‘true’ to indicate an invalid declaration.

In addition, set the source range for the annotation on the metaclass
identifier token to include the scope specifier only when a scope
specifier is present.