[Reflection] Parse metaclass definitions
Allow the declaration of metaclass members within the metaclass
definition (in the manner of a C++ class member specification).

Example:

  $class feline {
  public:
    feline();
    virtual ~feline() = default;

    unsigned int get_weight() const { return weight; }
    unsigned int get_kills() const { return kills; }

  private:
    unsigned int weight;
    unsigned int kills;
  }

Currently, metaclass definitions are parsed using
Parser::ParseCXXMemberSpecification(), and the members are stored in a
CXXRecordDecl variable within MetaclassDecl. This "nested" class shares
the same name as its enclosing metaclass, and is the only declaration
in its enclosing declaration context (the enclosing metaclass).

Metaclass members cannot be defined out-of-line.