[Parse][Sema][Reflection] Refactor ConstexprDecl semantic actions and improve parsing
Most notably, change the capture default in the synthesized lambda
expression from LCD_None to LCD_ByRef in order to be consistent with the
signature that is documented in Sema::ActOnConstexprDecl().

In Parser::ParseConstexprDeclaration(), if the parser encounters an
error, perform the necessary cleanups and do not evaluate the body.