Fixing #19 again.
The segfault was a result of an un-reflected declaration: an access
specifier (i.e., public:). Those are reflected as part of the AST.
This fix generates an "internal" reflection of access specifiers because
it's unlikely that we'll be able to support that in other implementations.
The "internal" nodes are filtered out of member sequences in the meta
library.