[Reflection] Treat class prototypes as “tag” declarations
Treat C++ class declarations that were declared using a metaclass-name
as the class-key as “tag” declarations and refer to them as “class
prototypes” in diagnostic messages.

Note: Class prototypes will still be (erroneously) handled like (or
referred to as) “structs” in other places, mainly because the
CXXRecordDecl objects that represent such declarations have a TagKind of
TTK_Struct (see TypeWithKeyword::getTagTypeKindForTypeSpec()).