[Reflection] Initial implementation of C++ “default” classes
As an alternative to the current implementation of C++ metaclasses,
extend class-head to allow the declaration of “default” classes (i.e.,
metaclasses) and class prototypes (prototype classes?) that are based on
a previously defined default class.

Example:

  class Ifoo default { /* … */ }; // Default class.
  class foo (Ifoo) { /* … */ }; // Class prototype.

In the declaration of a default class, the ‘default’ keyword must follow
the class-virt-specifier, if specified. Only structs and classes can be
default classes. Likewise, the default-specifier (the name of the
default class within parentheses after the class-key, which is named to
disambiguate it from the existing metaclass implementation) can only
appear on structs and classes.

Default class members are currently not injected into the class
prototype.