Fix instantiation bug in expansion statements.
Stop using a custom instantiator to manage local insantiations
when instantiating the body of the for loop.
Always check for local instantiations in FindInstantiation. It's
no longer the case that only a handful of declarations are local;
if a declaration has been instantiated locally, then it must be
found in that lookup.
Add an incomplete test for loop body instantiation.
Add specialized dump methods for new nodes.