Make injected default member initializers work.
Make sure we actually profile idexpr names correctly so they hash
to the right locations. This doesn't change the fact that idexpr
names can't be reliably used for lookup when they depend on e.g.,
local variables (that's a problem).

Automatically instantiate the default member initializer when
instantiating a field declaration.