Fixing nasty odr-use issues in expansion loops.
Statements and expressions within an expansion loop are not odr-used
unless they are instantiated. Or rather, those expressions are never
odr-used; the instantiated ones are. Make sure that Clang doesn't see
those loop bodies as odr-use contexts.

This is a new concept for Clang; non-odr-use contexts outside of
templates and expressions. This fix introduces a new framework that
can be used to identify when we're in an expansion loop body and
suppresses odr-use checking in those contexts.

Also, fix parameter numbering (again).

Make the pretty printer print actual types instead of reflected
types when they're available.