Working on instantiating non-trivial injected members.
Previous examples involving the injected class names were producing
members that linked to the wrong type. It turns out the previous
examples generated a class that could not be named or used within the
metaprogram, so we need a new approach.

Current work focuses on using the 'using class ... as ...;' facility
to inject new classes with lookupable names.

Also, fix a regression that tried to instantiate definitions of
undefined functions.