Redesign how fragments and injection work.
Definitions in class fragments are not parsed at the end of the
fragment; they will be parsed at the point of injection. Add features
to Sema and AST that allow that information to be propagated through
to injection.

Start decoupling injection from template substitution and rebuilding
the entire substitution structure.

Undo lookup changes based on previous ideas.