Initial support for checking type information.
The good NEWS is, we can reuse the functionality in SEMA
as it is available during analysis based warnings. This would
not be possible, and this might be a good argument to have the check
in the compiler rather than in tidy.

The bad news is, Analysis is the dependency of sema and not the other
way around. So in order to avoid circular dependendcy, we cannot
call directly into sema, rather we pass a callback.