Support ternary operator gnu extension
There is a gnu extension where a branch of the ternary operator
can be ommitted and it is assumed to be the same expression as the
condition. Fortunately, it is very easy to add support for that.

(cherry picked from commit 3b2c4334ceaa96cf201cca6fc6c14cc36dc83762)