Remove old lambda support and disable lambda related tests
As lambdas are no longer Pointers, we do not have the infrastructure to
track them yet.

I think the proper way to implement support is to handle them at the
point of assignment/binding. Whenever we bind a local variable to a
LambdaExpr, we could look up the capture list and explode the fields of
the local variable (adding them to the pmap).

(cherry picked from commit 5c0042bb6ee51beb90826acd49fa39d2dd2b781e)