Properly add the DerefLocations for references.
I do not really like the current design but we will need to refactor to
unify caller and callee analysis anyways. This is just a quick baseline
to make sure we do not regress until the conference.

An idea to make the voerall codebase cleaner:
Deref and prime are fundamentally similar. Maybe use deref method for
owners to add prime and represent the number of primes the same way as
derefs? The main difference between the two is the invalidation
semantics.

(cherry picked from commit 86cf7e1a1d1252af627403ce102acb46c8468408)